Implemented alternative derivation of Surprise. Although this version is congruent with the non-overlapping version of Surprise (in the case of non-overlapping communities), it still suffers from problems. In particular, duplicate communities will improve this measure (as long as they are above the average density of the graph it seems). Moreover, near duplicate communities are also abundant. This severely restricts the applicability of the method, and essentially renders it useless. For the moment, I leave this as is, perhaps we'll think of something still in the future.

+++ include/SurpriseVertexCover.h
@@ -19,7 +19,7 @@ class SurpriseVertexCover: public MutableVertexCover
     virtual double quality();
   protected:
   private:
-     double max_overlap; //Maximum overlap
+     size_t max_overlap; //Maximum overlap
 };
 
 #endif // SURPRISEVERTEXCover_H
